requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astroid==3.0.1
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cloudinary==1.38.0
colorama==0.4.6
coverage==7.3.2
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.4
django-appconf==1.0.6
django-ckeditor==6.7.0
django-environ==0.10.0
django-import-export==3.3.1
django-js-asset==2.1.0
django-paypal==2.0
et-xmlfile==1.1.0
filelock==3.16.0
gunicorn==21.2.0
identify==2.6.1
idna==3.4
isort==5.12.0
MarkupPy==1.14
mccabe==0.7.0
nodeenv==1.9.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.0
platformdirs==4.0.0
pre-commit==3.7.1
psycopg2==2.9.7
psycopg2-binary==2.9.9
pycparser==2.21
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.26.4
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
